requirements.txt
flask
anthropic
pypdf
python-docx
